requirements.txt
# Bot + Discord
nextcord

# Environment
python-dotenv

# FastAPI backend
fastapi
uvicorn[standard]

# Database
sqlalchemy
sqlalchemy[asyncio]
alembic
pymysql  # Required for SQLAlchemy + MySQL connector

# Validation
pydantic

# Optional utilities you already had
pyyaml
cryptography
